refactor(design-system): noun re-audit — ProductPageShell, ModelTable, BillingGateState
Three ruled component-hierarchy re-audit moves:

- ProductPageShell is a shape (props are ReactNode/string/boolean only),
  not a noun: moves from components/patterns/ into
  primitives/patterns/. Rewrites its 5 Workflow* importers.
- ModelTable fails the noun admission test (1 consumer, no domain-type
  import): demoted out of the library to sit beside its single call
  site, HarnessAllModelsSection, under
  components/settings/panes/agents/harness/.
- BillingGateState's locally declared view type (BillingGateKind,
  BillingGateAction, BillingGateStateView, BillingGateReason) and its
  presenter (billingGateView, toBillingGateReason) relocate to
  lib/domain/billing/billing-gate-presentation.ts, matching the
  pr-status-presentation.ts precedent. The rendering components
  (BillingGateState, BillingBalanceNotice) stay in
  components/patterns/BillingGateState.tsx and import the type from
  its new home.

Playground registry (components/playground/library/product-patterns.tsx),
its LIBRARY_TIERS placement, and DESIGN_SYSTEM.md's component index are
left for the integrator per the vocabulary-PR integration contract —
they reference all three old paths and will 404 until reconciled
against the other parallel re-audit moves.
