+++ apps/packages/product-client/src/components/patterns/BillingGateState.test.tsx
@@ -2,13 +2,8 @@
 import { describe, expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
 
-import {
-  BillingBalanceNotice,
-  BillingGateState,
-  billingGateView,
-  toBillingGateReason,
-  type BillingGateReason,
-} from "./BillingGateState";
+import { BillingBalanceNotice, BillingGateState } from "./BillingGateState";
+import { billingGateView } from "#product/lib/domain/billing/billing-gate-presentation";
 
 const noop = () => {};
 
@@ -19,137 +14,6 @@ const MANAGER = {
   onRefill: noop,
   onOpenBilling: noop,
 };
-
-describe("billingGateView", () => {
-  it("routes a free subject's exhausted credits to upgrade", () => {
-    const view = billingGateView("credits_exhausted", MANAGER);
-    expect(view.kind).toBe("upgrade");
-    expect(view.primaryAction?.label).toBe("Upgrade");
-    expect(view.secondaryAction?.label).toBe("Billing settings");
-  });
-
-  it("routes a paid subject's exhausted credits to refill", () => {
-    const view = billingGateView("credits_exhausted", { ...MANAGER, isPaidPlan: true });
-    expect(view.kind).toBe("refill");
-    expect(view.primaryAction?.label).toBe("Add credits");
-  });
-
-  it("gives non-admin members admin-managed copy and no action (U3: members cannot repair)", () => {
-    const view = billingGateView("credits_exhausted", {
-      ...MANAGER,
-      canManageBilling: false,
-    });
-    expect(view.kind).toBe("admin");
-    expect(view.primaryAction).toBeNull();
-    expect(view.description).toContain("organization admin");
-  });
-
-  it("keeps limit_reached distinct from exhausted: limits point at settings, never refill (N3)", () => {
-    const view = billingGateView("llm_limit_reached", { ...MANAGER, isPaidPlan: true });
-    expect(view.kind).toBe("limit");
-    expect(view.primaryAction?.label).toBe("Billing settings");
-  });
-
-  it("payment failure points at the portal path, not upgrade", () => {
-    const view = billingGateView("payment_failed", { ...MANAGER, isPaidPlan: true });
-    expect(view.kind).toBe("payment");
-    expect(view.primaryAction?.label).toBe("Billing settings");
-  });
-
-  it("admin hold offers no self-service repair", () => {
-    const view = billingGateView("admin_hold", MANAGER);
-    expect(view.primaryAction).toBeNull();
-    expect(view.secondaryAction?.label).toBe("Billing settings");
-  });
-
-  it("admin hold hides even the settings link from non-admin members", () => {
-    const view = billingGateView("admin_hold", { ...MANAGER, canManageBilling: false });
-    expect(view.primaryAction).toBeNull();
-    expect(view.secondaryAction).toBeNull();
-  });
-
-  it("concurrency limit is not a billing repair: no billing CTA", () => {
-    const view = billingGateView("concurrency_limit", MANAGER);
-    expect(view.title).toBe("Sandbox limit reached");
-    expect(view.primaryAction).toBeNull();
-    expect(view.secondaryAction).toBeNull();
-  });
-
-  it("overage_disabled and cap_exhausted point admins at settings only", () => {
-    for (const reason of ["overage_disabled", "cap_exhausted"] as const) {
-      const view = billingGateView(reason, { ...MANAGER, isPaidPlan: true });
-      expect(view.primaryAction?.label).toBe("Billing settings");
-      expect(view.secondaryAction).toBeNull();
-    }
-  });
-
-  it("external_billing_hold renders the payment path like payment_failed", () => {
-    const view = billingGateView("external_billing_hold", { ...MANAGER, isPaidPlan: true });
-    expect(view.kind).toBe("payment");
-    expect(view.primaryAction?.label).toBe("Billing settings");
-  });
-
-  it("never leaks a raw reason code: unknown reasons render the generic state", () => {
-    const view = billingGateView("unknown", MANAGER);
-    expect(view.title).toBe("Cloud usage paused");
-  });
-
-  it("falls back to billing settings when the surface has no checkout wiring", () => {
-    const view = billingGateView("credits_exhausted", {
-      isPaidPlan: false,
-      canManageBilling: true,
-      onOpenBilling: noop,
-    });
-    expect(view.primaryAction?.label).toBe("Billing settings");
-    expect(view.secondaryAction).toBeNull();
-  });
-
-  it("covers every typed reason without throwing", () => {
-    const reasons: BillingGateReason[] = [
-      "credits_exhausted",
-      "overage_disabled",
-      "cap_exhausted",
-      "payment_failed",
-      "external_billing_hold",
-      "admin_hold",
-      "concurrency_limit",
-      "llm_credits_exhausted",
-      "llm_limit_reached",
-      "unknown",
-    ];
-    for (const reason of reasons) {
-      const view = billingGateView(reason, MANAGER);
-      expect(view.title.length).toBeGreaterThan(0);
-      expect(view.description.length).toBeGreaterThan(0);
-    }
-  });
-});
-
-describe("toBillingGateReason", () => {
-  it("passes through every typed reason", () => {
-    const reasons: BillingGateReason[] = [
-      "credits_exhausted",
-      "overage_disabled",
-      "cap_exhausted",
-      "payment_failed",
-      "external_billing_hold",
-      "admin_hold",
-      "concurrency_limit",
-      "llm_credits_exhausted",
-      "llm_limit_reached",
-    ];
-    for (const reason of reasons) {
-      expect(toBillingGateReason(reason)).toBe(reason);
-    }
-  });
-
-  it("maps unrecognized, null, and empty codes to unknown", () => {
-    expect(toBillingGateReason("billing_quota")).toBe("unknown");
-    expect(toBillingGateReason(null)).toBe("unknown");
-    expect(toBillingGateReason(undefined)).toBe("unknown");
-    expect(toBillingGateReason("")).toBe("unknown");
-  });
-});
 
 describe("BillingGateState", () => {
   it("renders title, description, and both actions", () => {
