+++ apps/packages/product-client/src/components/settings/panes/agents/harness/HarnessAllModelsSection.tsx
@@ -15,7 +15,7 @@ import { AnimatedCollapsibleContent } from "#product/primitives/AnimatedCollapsi
 import { Button } from "#product/primitives/Button";
 import { IconButton } from "#product/primitives/IconButton";
 import { Input } from "#product/primitives/Input";
-import { ModelTable, type ModelTableRow } from "#product/components/patterns/ModelTable";
+import { ModelTable, type ModelTableRow } from "#product/components/settings/panes/agents/harness/ModelTable";
 import { HARNESS_PANE_COPY } from "#product/copy/settings/harness-pane";
 import { HarnessSection } from "#product/components/settings/panes/agents/harness/HarnessSection";
 import { useCloudAvailabilityState } from "#product/hooks/cloud/derived/use-cloud-availability-state";
