+++ apps/packages/product-client/src/primitives/patterns/ProductPageShell.tsx
